+++ Server/db_creation.sql
@@ -128,7 +128,7 @@ CREATE TABLE IF NOT EXISTS public.countries
     uuid uuid NOT NULL,
     name character varying COLLATE pg_catalog."default" NOT NULL,
     CONSTRAINT countries_pkey PRIMARY KEY (uuid)
-)
+);
 
 ALTER TABLE IF EXISTS public.city
     ADD COLUMN country_uuid uuid NOT NULL;
