+++ Server/package.json
@@ -4,7 +4,7 @@
   "description": "Search for local handyman to help you home",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --detectOpenHandles",
     "start": "node server.js",
     "watch": "nodemon server.js"
   },
@@ -20,5 +20,10 @@
     "swagger-jsdoc": "^6.2.5",
     "swagger-ui-express": "^4.5.0",
     "uuid": "^9.0.0"
+  },
+  "devDependencies": {
+    "@databases/pg-test": "^3.1.2",
+    "jest": "^29.2.1",
+    "supertest": "^6.3.0"
   }
 }
